Uebungen/26-web-service/src/BFF/requirements.txt
# tested with python 3.10.0
# upgrade pip
wheel
pyodbc
numpy
flask
flasgger
grpcio
grpcio-tools
